+++ packages/jazz-browser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # jazz-browser
 
+## 0.5.1
+
+### Patch Changes
+
+- Allow account migrations to be async
+- Updated dependencies
+  - cojson@0.5.2
+
 ## 0.5.0
 
 ### Minor Changes